Starting a new implementation in C++
This is the beggining of a new implementation of the program in C++. For now, I implemented a class named Monitor that can scan a Linux (and maybe other Unix like) system and find the files that allow to use the ACPI interface to control the backlight. The class has easy to use methods to change the backlight of a supported monitor.